refactor: 테스트 파일 패키지 분리
