+++ src/test/java/racing_car/CarTest.java
@@ -6,35 +6,17 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racing_car.model.Car;
 import racing_car.model.Cars;
-import racing_car.model.Game;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
 
-public class RacingCarTest {
-
-    private static Cars separateCarNamesByDelimiter(String target) {
-
-        return Cars.of(target.split(","));
-    }
-
-    @Test
-    @DisplayName("제공된 문자열로 자동차 배열 생성")
-    void separateByDelimiter() {
-        String testNames = "pobi,crong,honux";
-        Cars cars = separateCarNamesByDelimiter(testNames);
-        assertEquals(cars, Cars.of(new Car[]{
-                new Car("pobi"),
-                new Car("crong"),
-                new Car("honux")
-        }));
-    }
+public class CarTest {
 
     @Test
     @DisplayName("이름이 빈 문자열일 경우 테스트")
     void checkEmptyName() {
         String testName = "";
-        Throwable t = assertThrowsExactly(IllegalArgumentException.class, () -> separateCarNamesByDelimiter(testName));
+        Throwable t = assertThrowsExactly(IllegalArgumentException.class, () -> Cars.of(testName.split(",")));
         assertEquals(t.getMessage(), "이름은 1자 이상이여야 합니다.");
     }
 
@@ -42,7 +24,7 @@ public class RacingCarTest {
     @DisplayName("이름이 5자 초과일 경우 테스트")
     void checkNameMaxLength() {
         String testName = "chunsik";
-        Throwable t = assertThrowsExactly(IllegalArgumentException.class, () -> separateCarNamesByDelimiter(testName));
+        Throwable t = assertThrowsExactly(IllegalArgumentException.class, () -> Cars.of(testName.split(",")));
         assertEquals(t.getMessage(), "이름은 5자 이하여야 합니다.");
     }
 
@@ -67,24 +49,5 @@ public class RacingCarTest {
         b.move(2);
 
         assertEquals(-1, b.compareTo(a));
-    }
-
-    @Test
-    @DisplayName("우승자 구하기")
-    void getWinners() {
-        String[] testNames = {"pobi", "crong", "honux", "ryan", "chun", "jordy"};
-        Car[] cars = new Car[testNames.length];
-
-        int[] distances = new int[]{1, 5, 2, 5, 4, 5};
-        for (int i = 0; i < distances.length; i++) {
-            cars[i] = new Car(testNames[i]);
-            cars[i].move(distances[i]);
-        }
-
-        Game game = new Game(Cars.of(cars));
-
-        assertEquals(Cars.of(new Car[]{
-                cars[1], cars[3], cars[5]
-        }), game.getWinners());
     }
 }
